requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
Bottleneck @ file:///C:/miniconda3/conda-bld/bottleneck_1761938096667/work
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
idna==3.11
Jinja2==3.1.6
joblib @ file:///C:/miniconda3/conda-bld/joblib_1757926370033/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.46.0
MarkupSafe==3.0.3
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592981169/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593075610/work
narwhals==2.13.0
networkx==3.4.2
numba==0.63.1
numexpr @ file:///C:/miniconda3/conda-bld/numexpr_1762165723970/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1761579405538/work/dist/numpy-2.2.5-cp310-cp310-win_amd64.whl#sha256=074bffd27eaf24d54337a4bac76b167c7d794ddb6a813f6fa3fb37c91af766a8
packaging==25.0
pandas @ file:///C:/miniconda3/conda-bld/pandas_1762332399653/work/dist/pandas-2.3.3-cp310-cp310-win_amd64.whl#sha256=25368e0adc93a7db0f3983502b4e57344f71af8f45ca3849f7c3833738f306aa
pillow==12.0.0
protobuf==6.33.2
psutil @ file:///C:/miniconda3/conda-bld/psutil_1761896539676/work
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465513320/work
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn @ file:///C:/b/abs_45qemhn4lg/croot/scikit-learn_1753427401078/work
scipy @ file:///C:/miniconda3/conda-bld/scipy_1759222090244/work/dist/scipy-1.15.3-cp310-cp310-win_amd64.whl#sha256=2c02ce5f7bc108a9dd2298fac421e55898c14864219396421ffe00419ec8c736
shap==0.49.1
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
slicer==0.0.8
smmap==5.0.2
streamlit==1.52.1
tenacity==9.1.2
threadpoolctl @ file:///C:/miniconda3/conda-bld/threadpoolctl_1764361785970/work
toml==0.10.2
tornado==6.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.6.2
watchdog==6.0.0
